+++ src/gesturehandler/gesturehandler.android.ts
@@ -8,7 +8,7 @@ import {
     PanGestureHandlerOptions,
     PinchGestureHandlerOptions,
     RotationGestureHandlerOptions,
-    TapGestureHandlerOptions,
+    TapGestureHandlerOptions
 } from './gesturehandler';
 import {
     BaseGestureRootView,
@@ -27,7 +27,7 @@ import {
     ViewInitEvent,
     applyMixins,
     install as installBase,
-    nativeProperty,
+    nativeProperty
 } from './gesturehandler.common';
 import { observe as gestureObserve } from './gestures_override';
 
@@ -35,15 +35,15 @@ export { GestureState, GestureHandlerStateEvent, GestureHandlerTouchEvent, Gestu
 
 let PageLayout: typeof com.nativescript.gesturehandler.PageLayout;
 class PageGestureExtended extends Page {
-    nativeView: com.nativescript.gesturehandler.PageLayout;
+    nativeViewProtected: com.nativescript.gesturehandler.PageLayout;
     initNativeView() {
-        this.nativeView.initialize();
+        this.nativeViewProtected.initialize();
     }
     disposeNativeView() {
-        this.nativeView.tearDown();
+        this.nativeViewProtected.tearDown();
     }
     get registry() {
-        return this.nativeView && this.nativeView.registry();
+        return this.nativeViewProtected && this.nativeViewProtected.registry();
     }
 }
 let installed = false;
@@ -80,6 +80,21 @@ export function install(overrideNGestures = false) {
                 PageLayout = com.nativescript.gesturehandler.PageLayout;
             }
             const layout = new PageLayout(this._context, ROOT_GESTURE_HANDLER_TAG);
+
+            //@ts-ignore
+            if (layout.addRowsFromJSON) {
+                //@ts-ignore
+                layout.addRowsFromJSON(
+                    JSON.stringify([
+                        { value: 1, type: 0 /* org.nativescript.widgets.GridUnitType.auto */ },
+                        { value: 1, type: 2 /* org.nativescript.widgets.GridUnitType.star */ }
+                    ])
+                );
+            } else {
+                layout.addRow(new org.nativescript.widgets.ItemSpec(1, org.nativescript.widgets.GridUnitType.auto));
+                layout.addRow(new org.nativescript.widgets.ItemSpec(1, org.nativescript.widgets.GridUnitType.star));
+            }
+
             // this.gestureRegistry = layout.registry();
             return layout;
         };
@@ -195,9 +210,10 @@ export abstract class Handler<T extends com.swmansion.gesturehandler.GestureHand
                     }
                     return [left, top, right, bottom, width, height];
                 }
-            },
-        },
-    }) hitSlop;
+            }
+        }
+    })
+    hitSlop;
     @nativeProperty enabled: boolean;
     @nativeProperty shouldCancelWhenOutside: boolean;
     shouldStartGesture: (arg) => boolean;
@@ -216,7 +232,7 @@ export abstract class Handler<T extends com.swmansion.gesturehandler.GestureHand
             // x: Utils.layout.toDeviceIndependentPixels(handler.getX()),
             // y: Utils.layout.toDeviceIndependentPixels(handler.getY()),
             positions,
-            numberOfPointers,
+            numberOfPointers
         };
     }
     initNativeView(native: T, options: U) {
@@ -225,7 +241,7 @@ export abstract class Handler<T extends com.swmansion.gesturehandler.GestureHand
         this.touchListener = new com.swmansion.gesturehandler.OnTouchEventListener({
             shouldStartGesture: this.handleShouldStartGesture.bind(this),
             onTouchEvent: this.onTouchEvent.bind(this),
-            onStateChange: this.onStateChange.bind(this),
+            onStateChange: this.onStateChange.bind(this)
         });
         native.setOnTouchEventListener(this.touchListener);
         this.manager?.get()?.configureInteractions(this, options);
@@ -251,8 +267,8 @@ export abstract class Handler<T extends com.swmansion.gesturehandler.GestureHand
                 state: handler.getState(),
                 android: view,
                 extraData: this.getExtraData(handler),
-                view: view.nsView ? view.nsView?.get() : null,
-            },
+                view: view.nsView ? view.nsView?.get() : null
+            }
         });
     }
     onStateChange(handler: T, state: number, prevState: number) {
@@ -265,8 +281,8 @@ export abstract class Handler<T extends com.swmansion.gesturehandler.GestureHand
                 prevState,
                 android: view,
                 extraData: this.getExtraData(handler),
-                view: view.nsView ? view.nsView?.get() : null,
-            },
+                view: view.nsView ? view.nsView?.get() : null
+            }
         });
     }
 
@@ -327,7 +343,7 @@ export class TapGestureHandler extends Handler<com.swmansion.gesturehandler.TapG
             x: Utils.layout.toDeviceIndependentPixels(handler.getLastRelativePositionX()),
             y: Utils.layout.toDeviceIndependentPixels(handler.getLastRelativePositionY()),
             absoluteX: Utils.layout.toDeviceIndependentPixels(handler.getLastAbsolutePositionX()),
-            absoluteY: Utils.layout.toDeviceIndependentPixels(handler.getLastAbsolutePositionY()),
+            absoluteY: Utils.layout.toDeviceIndependentPixels(handler.getLastAbsolutePositionY())
         });
     }
 }
@@ -350,7 +366,7 @@ export class PanGestureHandler extends Handler<com.swmansion.gesturehandler.PanG
     @nativeProperty({ nativeSetterName: 'setAverageTouches' }) avgTouches: number;
     @nativeProperty numberOfPointers: number;
     createNative(options) {
-        const context = Application.android.context as android.content.Context;
+        const context = Utils.android.getApplicationContext() as android.content.Context;
         return new com.swmansion.gesturehandler.PanGestureHandler(context);
     }
     getExtraData(handler: com.swmansion.gesturehandler.PanGestureHandler) {
@@ -362,7 +378,7 @@ export class PanGestureHandler extends Handler<com.swmansion.gesturehandler.PanG
             translationX: Utils.layout.toDeviceIndependentPixels(handler.getTranslationX()),
             translationY: Utils.layout.toDeviceIndependentPixels(handler.getTranslationY()),
             velocityX: Utils.layout.toDeviceIndependentPixels(handler.getVelocityX()),
-            velocityY: Utils.layout.toDeviceIndependentPixels(handler.getVelocityY()),
+            velocityY: Utils.layout.toDeviceIndependentPixels(handler.getVelocityY())
         });
     }
 }
@@ -381,7 +397,7 @@ export class PinchGestureHandler extends Handler<com.swmansion.gesturehandler.Pi
             scale: handler.getScale(),
             focalX: Utils.layout.toDeviceIndependentPixels(handler.getFocalPointX()),
             focalY: Utils.layout.toDeviceIndependentPixels(handler.getFocalPointY()),
-            velocity: handler.getVelocity(),
+            velocity: handler.getVelocity()
         });
     }
 }
@@ -391,7 +407,7 @@ export enum FlingDirection {
     DIRECTION_LEFT = GestureHandler.DIRECTION_LEFT,
     DIRECTION_UP = GestureHandler.DIRECTION_UP,
     DIRECTION_DOWN = GestureHandler.DIRECTION_DOWN,
-    DIRECTION_RIGHT = GestureHandler.DIRECTION_RIGHT,
+    DIRECTION_RIGHT = GestureHandler.DIRECTION_RIGHT
 }
 
 function directionToString(direction: number) {
@@ -428,7 +444,7 @@ export class FlingGestureHandler extends Handler<com.swmansion.gesturehandler.Fl
     }
     getExtraData(handler: com.swmansion.gesturehandler.FlingGestureHandler) {
         return Object.assign(super.getExtraData(handler), {
-            direction: directionToString(handler.getRecognizedDirection()),
+            direction: directionToString(handler.getRecognizedDirection())
         });
     }
 }
@@ -436,7 +452,7 @@ export class LongPressGestureHandler extends Handler<com.swmansion.gesturehandle
     @nativeProperty minDurationMs: number;
     @nativeProperty({ converter: { fromNative: Utils.layout.toDevicePixels } }) maxDist: number;
     createNative(options) {
-        const context = Application.android.context as android.content.Context;
+        const context = Utils.android.getApplicationContext() as android.content.Context;
         return new com.swmansion.gesturehandler.LongPressGestureHandler(context);
     }
     getExtraData(handler: com.swmansion.gesturehandler.LongPressGestureHandler) {
@@ -444,7 +460,7 @@ export class LongPressGestureHandler extends Handler<com.swmansion.gesturehandle
             x: Utils.layout.toDeviceIndependentPixels(handler.getLastRelativePositionX()),
             y: Utils.layout.toDeviceIndependentPixels(handler.getLastRelativePositionY()),
             absoluteX: Utils.layout.toDeviceIndependentPixels(handler.getLastAbsolutePositionX()),
-            absoluteY: Utils.layout.toDeviceIndependentPixels(handler.getLastAbsolutePositionY()),
+            absoluteY: Utils.layout.toDeviceIndependentPixels(handler.getLastAbsolutePositionY())
         });
     }
 }
@@ -461,7 +477,7 @@ export class RotationGestureHandler extends Handler<com.swmansion.gesturehandler
             rotation: handler.getRotation(),
             anchorX: Utils.layout.toDeviceIndependentPixels(handler.getAnchorX()),
             anchorY: Utils.layout.toDeviceIndependentPixels(handler.getAnchorY()),
-            velocity: handler.getVelocity(),
+            velocity: handler.getVelocity()
         });
     }
 }
@@ -474,7 +490,7 @@ export class NativeViewGestureHandler extends Handler<com.swmansion.gesturehandl
     }
     getExtraData(handler: com.swmansion.gesturehandler.NativeViewGestureHandler) {
         return Object.assign(super.getExtraData(handler), {
-            pointerInside: handler.isWithinBounds(),
+            pointerInside: handler.isWithinBounds()
         });
     }
 }
@@ -586,7 +602,7 @@ export class Manager extends ManagerBase {
         }
         viewListeners.set(handler.getTag(), {
             init: onInit,
-            dispose: onDispose,
+            dispose: onDispose
         });
     }
     detachGestureHandler<T extends com.swmansion.gesturehandler.GestureHandler<any> = com.swmansion.gesturehandler.GestureHandler<any>>(handler: Handler<T, any>, view: View) {